Add function to get user by ID
This is needed when simulating receiving encrypted messages from remote clients: we want to let mocktransportsession handle all non-self client encryption contexts. So we will need to create a remote client for each local client (that is not self), and retrieve them later when we want to locally simulate and encrypted messages.